mkbootfs: Fix the default st_mode for root directory.
CL in [1] fixed the wrong mode for the root directory from 0644 to 0755.
However it only handles the path by calling mkbootfs with canned
fs_config (i.e. with -f flag). When calling mkbootfs without -f flag, it
still treats the root directory as a file. This renders the generated
boot/recovery images w/ and w/o -f being different, and leads to the
incorrect checksums in /system/bin/install-recovery.sh.

[1] commit aa8f2f65a030f71506277e2a8d64d83a099e9feb.

Bug: 31988535
Test: `make dist` and verify the checksums in /system/bin/install-recovery.sh.

Change-Id: Ib50fadb23367da0b46944ae0579093da21412593